Add fuzzer test for whole staking activation flow
